Creating a log interface
This is designed to replace winston through the system with a focus on request oriented logging. This is setup to by default push all messages through rabbitmq but allows any number of transports. This is somewhat reinventing the wheel but I think it'll be more effective. I don't like logging libraries...